requirements.txt
# Bot Performance and Memory Optimization Requirements

# Core Discord bot requirements
discord.py>=2.3.0
python-dotenv>=1.0.0


# Performance monitoring and system info
psutil>=5.9.0
aiohttp>=3.8.0
dnspython>=2.3.0

# Note: asyncio, weakref, threading, collections, hashlib, pickle, dataclasses, functools
# are all part of Python's standard library and don't need to be installed via pip
